mkimage: sunxi_egon: Allow overriding the padding size
Due to a bug in the H3 SoC, where the CPU 0 hotplug flag cannot be
written, resuming CPU 0 requires using the "Super Standby" code path in
the BROM instead of the hotplug path. This path requires jumping to an
eGON image in SRAM.

This resume image, whose only purpose is to jump back to the secure
monitor, contains a single instruction. Padding the image to 8 KiB would
be wasteful of SRAM. Hook up the -B (block size) option so users can set
the block/padding size.

Signed-off-by: Samuel Holland <samuel@sholland.org>