tools: Separate image types which depend on OpenSSL
Some image types always depend on OpenSSL, so they can only be included
in mkimage when TOOLS_LIBCRYPTO is selected. Instead of using arbitrary
preprocessor definitions, conditionally link the files.

When building for platforms which use those image types, automatically
select TOOLS_LIBCRYPTO, it is required for the build to complete.

Series-changes: 2
 - Refactored the first patch on top of TOOLS_LIBCRYPTO

Signed-off-by: Samuel Holland <samuel@sholland.org>