clk: sunxi: Move header out of arch directory
The CCU header is only used by the DM drivers, not any platform code.
Its current location adds an artificial dependency on CONFIG_ARM and
ARCH_SUNXI, which will be problematic when adding the CCU driver for
a RISC-V sunxi platform.

Commit-notes:
This patch trivially conflicts with the one adding the PRCM drivers.
Feel free to fix that when applying, or I can rebase either patch.
END
Series-to: sunxi
Signed-off-by: Samuel Holland <samuel@sholland.org>