sunxi: gpio: Remove name_to_gpio macro
This clarifies which callers must be updated to complete the DM_GPIO
conversion.

The only caller of name_to_gpio in generic code is inside the !DM_GPIO
block in cmd/gpio.c. DM_GPIO is always selected on sunxi, so that code
cannot be reached. And after this commit, there are only two remaining
implementations of name_to_gpio.

Signed-off-by: Samuel Holland <samuel@sholland.org>