sunxi: gpio: Remove bank-specific size macros
Since the beginning, all banks have had space for 32 pins, even when
not all pins were implemented. Let's use a single constant for the GPIO
bank size here, like the GPIO driver is already doing.

Cover-letter:
sunxi: Cleanup to prepare for DM_GPIO/DM_PINCTRL
I have a patch series in progresss to migrate sunxi to a DM_PINCTRL
driver and convert the remaining legacy GPIO users (outside of SPL).

This is one of several small independent sets of preparatory patches
I will be sending in order to minimize the size of the main series
(which has >100 patches at this point).
END
Series-to: sunxi
Signed-off-by: Samuel Holland <samuel@sholland.org>