phy: sun4i-usb: Use DM_GPIO for id_det GPIO
Now that the sunxi_gpio driver handles pull-up/down via the driver
model, we can switch to DM_GPIO for this pin with no loss in
functionality. Since this now gets the pin configuration from the
device tree, we can remove the Kconfig symbol.

The other pins rely on the special treatment of fake AXP PMIC GPIOs
in sunxi_name_to_gpio(), so they cannot be converted yet.

Signed-off-by: Samuel Holland <samuel@sholland.org>