phy: sun4i-usb: Support VBUS detection via power supply
The device tree binding provides for getting VBUS state from a device
referenced by phandle, as an optional alternative to using a GPIO. In
U-Boot, where there is no power supply class, this VBUS detection will
be implemented using a regulator device and its get_enable method.
Let's hook this up to the PHY driver.

Cover-letter:
phy: sun4i-usb: Support VBUS detection via power supply
This converts a hack (a virtual GPIO pin) to less of a hack (not-really
a regulator), but at least it now supports the device tree. This series
adds the consumer side. The new code won't do anything useful until the
regulator driver is merged (which depends on the DM_PMIC driver), but it
shouldn't hurt anything either.
END
Series-to: sunxi

Signed-off-by: Samuel Holland <samuel@sholland.org>