i2c: sun8i_rsb: Add support for DM clocks and resets
Currently, clock/reset setup for this device is handled by a
platform-specific function and is intermixed with non-DM pinctrl
setup. Use the devicetree to get clocks/resets, which disentagles
it from the pinctrl setup in preparation for moving to DM_PINCTRL.

This also has the added benefit of picking the right clock/reset
bits for H6 and new SoCs that have a rearranged PRCM MMIO space.

Cover-letter:
sunxi: Fixes for DM I2C drivers
A while back, the sunxi-specific P2WI and RSB drivers were converted to
support DM_I2C. Now they are used as non-DM drivers in SPL, and DM
drivers in U-Boot proper.

However, the DM version of the code did not fully initialize either the
controller or the connected chips. So the DM driver would only work if
the non-DM version had previously been used in SPL.

With these bug fixes and the pinctrl series, the drivers now work on
SoCs like A64 and H6, which have a PMIC but do not set it up in SPL.
END

Series-to: Heiko Schocher <hs@denx.de>
Series-to: sunxi
Series-cc: u-boot@lists.denx.de

Signed-off-by: Samuel Holland <samuel@sholland.org>